requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
requests>=2.32.0
PyMySQL>=1.1.0
aiohttp>=3.10.0
tesla-fleet-api>=1.4.5
openai>=1.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
